bereite_zu Methode wurde implementiert

+++ Cocktailpro/Mischer.cpp
@@ -15,7 +15,7 @@ Mischer::~Mischer(){
 /**
  * 
  */
-void Mischer::mischen(int zeit)
+void Mischer::mischen(float zeit)
 {
 	
 }
